Mail vs Router decisions stable, putting on polish and error checking

+++ Project_scripts/base/base_configuration.sh
@@ -20,10 +20,12 @@ set -o nounset                              # Treat unset variables as an erro
 function dobase {
 
 	source ./base/basevar
-
-	#sethostname
-	 hostnamectl set-hostname $hostnm
-
+		#sethostname
+	if [ $# -eq 0 ]; then
+		hostnamectl set-hostname $hostnm
+	else
+	 hostnamectl set-hostname $mailhostnm
+  fi
 	 echo "++++++++ firewalld is off ++++++++"
 	 systemctl disable firewalld
 	 systemctl stop firewalld
@@ -31,9 +33,14 @@ function dobase {
 
 echo "group_package_types=mandatory,default,optional" >> /etc/yum.conf
 		yum -y group install base
-		#  install EPEL software repository
+		if [ $? != 0 ]; then
+			echo "could not resolve DNS or no network connectivity please confirm you have internet then rerun this script "
+				exit 1
+		fi
+		#install EPEL software repository
 		echo "++++++ Installing EPEL +++++++"
 		yum -y install epel-release
+
 echo " "
 	echo "++++++ Updating Entire System +++++++"
  	yum -y update
